pyproject.toml
[project]
name = "week4"
version = "1.0.0"
description = "Paralympics testing activities"
dependencies = [
    "flask",
    "Flask-SQLAlchemy",
    "Flask-Marshmallow",
    "marshmallow-sqlalchemy",
    "pandas",
    "pytest"
]

# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["paralympics*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

